In the actus module: 1- Moved the deploy_contract test to initialize_pam. Makes more sense there. The deploy_contract function just outsources all work to initialize_pam. 2- Started writing tests for the year_fraction function.
